training docs + images translated in IT (italian)
I kept the same exact structure and tried to fit 100% of meaning with the translation. I left some original words where here in Italy is better to us them as they are (i.e. "overhead").
For the folder structure, I've just created an "it" subfolder waiting to rearrange them according to the web framework or design requirements